+++ src/extension/httpyacExtensionApi.ts
@@ -1,25 +1,22 @@
 import * as vscode from 'vscode';
-import * as httpyac from 'httpyac';
-
-export interface ResponseHandlerResult {
-  document: vscode.TextDocument;
-  editor: vscode.TextEditor;
-  uri?: vscode.Uri;
-}
+import type * as httpyac from 'httpyac';
 
 export interface ResponseItem {
-  created: Date;
-  name: string;
-  response: httpyac.HttpResponse;
-  httpRegion?: httpyac.HttpRegion;
-  document?: vscode.TextDocument;
-  uri?: vscode.Uri;
+  readonly id: string;
+  readonly created: Date;
+  readonly name: string;
+  readonly openWith?: string;
+  readonly extension: string;
+  readonly testResults?: Array<httpyac.TestResult>;
+  readonly metaData: Record<string, unknown>;
+  readonly response: httpyac.HttpResponse;
+  responseUri?: vscode.Uri;
+  documentUri?: vscode.Uri;
+  isCachedResponse: boolean;
+  loadResponseBody?(): Promise<void>;
 }
 
-export type ResponseHandler = (
-  response: httpyac.HttpResponse,
-  httpRegion?: httpyac.HttpRegion
-) => Promise<boolean | ResponseHandlerResult>;
+export type ResponseHandler = (responseItem: ResponseItem) => Promise<boolean>;
 
 export interface ResponseOutputProcessor {
   show: httpyac.RequestLogger;
@@ -29,6 +26,8 @@ export interface DocumentStore {
   readonly documentStoreChangedEmitter: vscode.EventEmitter<void>;
   readonly httpFileStore: httpyac.store.HttpFileStore;
   activeEnvironment: Array<string> | undefined;
+  getActiveEnvironment(httpFile: httpyac.HttpFile): Array<string> | undefined;
+  setActiveEnvironment(httpFile: httpyac.HttpFile, env: Array<string> | undefined): void;
   getDocumentPathLike: (document: vscode.TextDocument) => httpyac.PathLike;
   getHttpFile(document: vscode.TextDocument): Promise<httpyac.HttpFile | undefined>;
   getAll(): Array<httpyac.HttpFile>;
@@ -36,13 +35,15 @@ export interface DocumentStore {
   parse(uri: vscode.Uri | undefined, text: string): Promise<httpyac.HttpFile>;
   remove(document: vscode.TextDocument): void;
   send: (context: httpyac.HttpFileSendContext | httpyac.HttpRegionsSendContext) => Promise<boolean>;
+  clear(): void;
 }
 
 export interface ResponseStore {
+  readonly hasItems: boolean;
   readonly historyChanged: vscode.Event<void>;
   add(response: httpyac.HttpResponse, httpRegion?: httpyac.HttpRegion, show?: boolean): Promise<void>;
-  remove(responseItem: ResponseItem): boolean;
-  clear(): void;
+  remove(responseItem: ResponseItem): Promise<boolean>;
+  clear(): Promise<void>;
 }
 
 export interface HttpYacExtensionApi {
@@ -50,6 +51,7 @@ export interface HttpYacExtensionApi {
   documentStore: DocumentStore;
   responseStore: ResponseStore;
   httpDocumentSelector: vscode.DocumentSelector;
+  allHttpDocumentSelector: vscode.DocumentSelector;
   environmentChanged: vscode.Event<string[] | undefined>;
   getEnvironmentConfig(path: httpyac.PathLike): Promise<httpyac.EnvironmentConfig>;
   getErrorQuickFix: (err: Error) => string | undefined;